+++ include/RenderDX.h
@@ -61,7 +61,7 @@ namespace Easy3D {
         //BO
 		//virtual BaseVertexBufferObject* createVBO(const byte* vbo, size_t size, size_t strip, size_t offset)=0;
 		virtual BaseVertexBufferObject* createVBO(const byte* vbo, size_t stride, size_t n);
-		virtual BaseIndextBufferObject* createIBO(const ushort* ibo, size_t size);
+		virtual BaseIndextBufferObject* createIBO(const uint* ibo, size_t size);
         
 		virtual void bindVBO(BaseVertexBufferObject*);
 		virtual void bindIBO(BaseIndextBufferObject*);
